apps/api/requirements/base.txt
# Core Django dependencies
Django>=5.2,<5.3
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0

# Configuration
django-environ>=0.11,<1.0

# Security and middleware
django-cors-headers>=4.3,<5.0
argon2-cffi>=23.1,<24.0
whitenoise>=6.6,<7.0

# Production server
gunicorn>=22.0,<23.0

# Caching and async tasks
redis>=5.0,<6.0
celery>=5.4,<6.0

# Image processing
django-imagekit>=5.0,<6.0
Pillow>=10.0,<11.0

# Nested serializers
drf-writable-nested>=0.7.0,<1.0

# PDF generation (receipts)
weasyprint>=68.0,<69.0

# QR code generation
segno>=1.6.6,<2.0

# WebSocket support (Django Channels)
channels>=4.3.2,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0

# Audit history tracking
django-simple-history>=3.11.0,<4.0

# Finite state machine for payment/delivery status
# Using django-fsm-2 (maintained fork of django-fsm)
django-fsm-2>=3.0,<5.0

# GeoDjango support for delivery zones
djangorestframework-gis>=1.1.0,<2.0
